LibWeb: Check if `navigationParams` is a navigationParams struct
In the case of a network error (e.g. file does not exist), navigationParams
will be a `AK::StringView` containing an error message. The spec does
not consider this to be a possible value for navigationParams, so we
have to add an additional check before making an assertion that
navigationParams's request is not null.

This fixes an issue where the verification was failing, and also fixes
the issue where the error page would not be navigated if a file:// URL
does not exist.